ComputeSalaryHA3
Develop an abstract class Staff with ID, name as attributes and an abstract method compute Salary():double. The classes TeachingStaff and NonTeachingStaff inherit Staff. NonTeachingStaff has hoursWorked and hourlyWage as its attributes and the TeachingStaff has basic and allowance as attributes. The rules for the computation of salary are as follows: 1. For NonTeachingStaff, the salary is product of hoursWorked and hourlyWage. 2. For TeachingStaff the salary is basic + HRA(20 % of basic)+ allowance. Code the constructors, toString() method.Write the Demo class with main() method to create objects of both subclasses and compute the salary.